Remove profile blurb, add to sidebar instead
